tools: fixed port arg to serve_folder
git-svn-id: https://dev.lagrummet.se/svn/rinfo/trunk@1204 f32dfbcc-ef79-4080-89f5-1451246e9e1e

+++ tools/rinfomain/serve_folder.groovy
@@ -16,6 +16,6 @@ def serveDirectory(port, wwwDir) {
 }
 
 def wwwDir = args[0]
-def port = args.length > 1? args[1] : 8280
+def port = args.length > 1? args[1] as Integer : 8280
 serveDirectory(port, new File(wwwDir).toURI().toString())
